Ref #19: Implement ApproveProducer
An account can now approve a producer, and that producer's votes are
increased appropriately. The account can remove approval, and the votes
are decreased appropriately. Additionally, the producer is added/removed
to the account's approved producers list as desired.

TODO: A producer with more votes than the rest is still never rotated in