Merge apply_staked_claim into apply_eos_claim
The apply_staked_claim function appears to be defunct, a leftover of a
bygone era. It's logic, however, was not included in the still-active
apply_eos_claim, and that is a problem.

Move that missing logic into apply_eos_claim, pretty it up a bit, and
get rid of the vestigial apply_staked_claim.