Ref #203: Bugfix
Fix a slippery little bug in the new eosc, which was causing messages'
data payloads to be a serialized Bytes object rather than merely the
bytes contained by the object. The difference was just one length byte
at the beginning of the message data payload, but of course that was
enough to break parsing.