Add plugin for generating test transactions inside of eosd
txn_test_gen_plugin is a plugin that can generate test transactions at a configurable rate in transactions/sec. Since the transactions are generated internal to eosd, there is very little overhead compared to generating the transactions via eosc or other RPC mechanisms. This current implementation is a little rough; only having a rather coarse adjustment for transactions/sec and limiting it to a maximum of 400.