Correct WAVM table index trap
There looks to be an error on how the WAVM code traps table execution on elements beyond maximum table size. It compares the index in to the table to make sure it is less than the table size. But, the table is 16bytes per index -- so the number of bytes in the table is 16 times more than the number of eelements it can hold.

This commit fixes this error. I've confirmed that the LLVM optimizer still removes the entire check -- after all, any UINT32 value can never be greater than or equal to 2^32