More explicit block merkle tree calculation
Previously, the merkle tree was calculated across the input_transactions of a block. But this is a bit troublesome with block summaries since they do not contain input_transactions. While one option would be to record the ordering of input_transactions in to a block summary, instead we can define the merkle on top of the regions/cycles/sumarries that already exist in a summary and not need any redundant ordering indicators in the summary.