Fixup two recent unit tests for new sbrk behavior
Two recently fixed unit tests have a problem with new sbrk behavior. One simply needs modification to look for a -1 indicating failure. The other one, where sbrk and malloc is mixed, is more troublesome. When I adjusted the malloc behavior to only create heaps on wasm page boundaries it breaks that unit test. However, I prefer to keep the heaps on the wasm boundaries. So I’m going to disable that unit test for now,