Increase max wasm (smart contract) memory usage to 33MiB
Change the maximum memory a wasm smart contract can consume to 33MiB. Update a couple of the unit tests to accommodate any max memory size. A number of the “legacy” unit tests assume a 1MiB maximum so add a way in our cmakes to bound certain wasms to a particular runtime limit.