make get_table_rows via chain_plugin (and thus cleos) easier to use:
Added `symbol_name` built_in_type to abi_serializer, which is like
`symbol` except it does not include the leading precision in the
displayed string.

Now treating scope string as (in order of attempts): `name`, `uint64_t`,
`symbol`, or `symbol_name`. Only throw error if the string cannot be
interpretted as any of those four types. Ambiguity between a name
consisting of only the digits 1 through 5 and an actual `uint64_t`
number can be resolved by adding whitespace to the beginning or end of
the string representation of the `uint64_t` number passed into cleos.

Switched order of scope and code arguments in cleos to better reflect
the table structure.

Updated currency.abi to use the more descriptive types (such as `symbol`
and `symbol_name`) and also to use the correct table name ("accounts")
for the accounts tables.