intrinsics to return current and published time now return microseconds
The now() intrinsic was replaced with a current_time() intrinsic which returns the current time (not head block time) in microseconds since epoch.

A now() function was defined in eosiolib/system.h which returns current_time()/1000000 so that existing contracts which use now() need minimal changes.

The publication_time() intrinsic was modified to also return the time as microseconds since epoch.
The only contracts that used publication_time() so far were for tests and those tests have been updated to reflect the new meaning of publication_time() and current_time().