Make the default key added to each wallet optional
By default we add the "root" eosio key to each created wallet. This key is configurable, however it also is always required. Change to code to accept an empty key which is an indicator that no key shall be added to created wallets.

Orginally did this because it makes an incoming unit test a little easier; but I can also see why some users may not way a key placed in every single wallet they create too.