Support -r/--header option to pass specific HTTP header
It's common that 'cleos' and 'nodeos' are deployed separately into
different locations and they communicate across a network. In such case,
it's important to let 'cleos' has the ability to include HTTP headers so
that it can go through some 'middle' devices which requires some kind of
'token' included in the requests.

So a new option for 'cleos' in included to provide this ability.