zero wasm linear memory on memory growth
When a WASM grows its linear memory (via the grow_memory opcode) the new freshly accessible memory is required to be zeroed out.

Our binaryen implementation needs a fix for this.

WAVM is more interesting. It relies on a madvise(MADV_DONTNEED) to mark pages such that a subsequent access to those pages would be zero. This works on Linux.. but does not on macOS. It doesn't appear like this zero behavior that Linux performs is anything to do with what POSIX allows so macOS is probably within its rights to not zero memory.

Change the WAVM memory growth to simply memset() on growth. This is a little undesirable because it means the memory will get zeroed twice on Linux (once in the memset and once in the kernel due to the madvise)