wallet_api interface & soft_wallet refactor
This is a non-functionally change that creates a wallet_api interface and renames the current wallet class to soft_wallet. wallet_manager will now operate strictly on wallet_api instances with the exception of when it is asked to create/open a wallet in which case it will create a soft_wallet instance.

The idea is that hardware wallet implementations will implement wallet_api. wallet_manager will add instances of these hardware wallet objects to its wallet_api list.

A couple other minor changes:

Addition of a list_public_keys() call and wallet_manager now using it for its get_public_keys() call. Hardware wallets won't have the ability to get private keys, so a function that didn't return a private_key_type was required.

Addition of a is_lockable() call. Some wallet_api implementations for hardware wallets may not have the ability to lock. At least in their initial implementation. I would like attempt to lock unlockable wallets to fail; so add a way for lock_all() to query if it should lock before attempting to lock said wallets. This limitation could be removed in the future if lockablity is better understood in all cases.