remove speculation when stale
in the current default mode, a node attempts to act as a good p2p citizen and relay transactions.  This requires creating a speculative block.  However, in times when a network has steady transactions and the node is stale, this creates a storm of database activity managing our transaction deduplication list as "time" bounces from current for speculation to stale for applying incoming blocks.

The new default is to only open a speculative block when the node is in sync.  If transactions come in while the node is out of sync they are queued instead of immidiately executed and relayed.  The result is a significant speed up when syncing in to a live network with some traffic on it.

The old default can be restored by using the new `enable-stale-speculation` config parameter