(appbase sync) Provide a version string based off "git describe"
Appbase will now provide a version string that is based off "git describe --tags --dirty". This provides a readable version number like v1.1.0 as well as information on intermediate git revisions and/or dirtiness. Since we don’t provide eosio as a .tar.gz package due to Github’s lack of submodules on tarballing, generally all users should be building in a git clone.

Be aware this change will run git describe on EVERY build. The command is quite fast (around 100ms even on a slower 2017 MacBook, but sub 30ms on a quality Linux dev box) and will hopefully provide enough benefit to be worth it.