only fetch unapplied transactions if we need them
In sync and real-time scenarios under moderate to heavy traffic load there can be a large backlog of unapplied transactions.  Fetching the unapplied transactions has a cost associated due to the container conversion implied.  However, that map of transactions is only useful if we are:

* an RPC node that wants to persist incoming transactions state mutations until they get into a block or expire and need to determine the intersection of the persisted and unapplied transactions
* OR we are a producer trying to apply transactions learned but not yet applied

In all other cases, we can avoid the cost of fetching the unapplied transactions outright