fix dereference empty optional bug in api_tests/transaction_tests
The applied_transaction signal can be given a transaction_trace for a 
failed input transaction in which the receipt optional is empty.

Disconnect the connection to the applied_transaction signal that was 
only meant for the send_transaction_trigger_error_handler test after 
that particular unit test is completed.
This ensures that the stale signal handling lambda function is not 
called when applying the later transactions in the test.

Furthermore, the signal handler should still check if the receipt is 
empty before attempting to dereference it even for the 
send_transaction_trigger_error_handler unit test, since the initial 
transaction to send the deferred transaction could fail, and it is 
better for the official failure in the Boost tests to be from 
BOOST_REQUIRE(trace) rather than from a SIGABRT.