Fix bug that creates a gap in the syncing blocks
When transitioning from the LIB sync mode to the mode where net_plugin catches up to the head of a remote peers best known branch, we calculate the branch to send based on the peers local LIB which could have shifted since the time the request was made due to latency.  If this it the first time we are synchronizing with that branch this means that there could be a gap in delivered blocks resulting in the blocks sent being unlinkable.   The more latency on a connection the more likely it is that the LIB of your block provider moved forward between the packets.

As a fix, we now send, at most, the branch from the LIB we last sent to that peer instead of the current LIB.  In addition, if that peer has reported its head id and we are building on top of that branch, we also squelch the delivery of blocks they already know about.